creación de componente para el login, y modificación de la ruta detalle
