+++ src/pages/Index.js
@@ -6,6 +6,7 @@ export { default as Us } from './Us/Us';
 export { default as Profile } from './Profile/Profile';
 export { default as NotFound } from './NotFound/NotFound';
 export { default as Products } from './Products/Products';
+export { default as Orders } from './Orders/Orders';
 // export { default as Favorite } from './Favorite/Favorite';
 // Pages Admin
 // export { default as Category } from './Admin/Category';
