+++ src/pages/Profile/Profile.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import { useAuth } from "../../hooks/useAuth";
 import { CustomAlert, RegisterForm } from "../../components";
